Fixed bug with LCD address (truncated to uint16_t).
git-svn-id: http://svn.tilp.info/repos/tiemu@742 4f589446-7ac1-0310-82c7-a055b08196ef

+++ tiemu/trunk/src/core/ti68k_def.h
@@ -120,7 +120,7 @@ typedef struct
     int     on_key;
 
     // lcd.c
-	uint16_t lcd_adr;		// LCD address (as $4c00)
+	uint32_t lcd_adr;		// LCD address (as $4c00)
 	char	*lcd_ptr;		// direct pointer to LCD in PC RAM
     int     contrast;
 	int		log_w;			// LCD logical width
